backend/mock/requirements.txt
Django>=5.1
psycopg2-binary>=2.9
python-dotenv>=1.0.1
gunicorn>=23.0.0
djangorestframework>=3.15.2
strawberry-graphql-django>=0.57.1
strawberry-django-auth>=0.378.4
django-cors-headers>=4.7.0
requests>=2.32.3
cryptography>=44.0.2
python-dotenv>=1.0.1
django-code-generator